% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_generator.R
\name{CurveGenerator}
\alias{CurveGenerator}
\title{CurveGenerator Class}
\description{
Base class that provides a common structure to the logistic and dome approach to logistic classes
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lengths}}{vector of length classes}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CurveGenerator-new}{\code{CurveGenerator$new()}}
\item \href{#method-CurveGenerator-run}{\code{CurveGenerator$run()}}
\item \href{#method-CurveGenerator-generate_curve_df}{\code{CurveGenerator$generate_curve_df()}}
\item \href{#method-CurveGenerator-clone}{\code{CurveGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CurveGenerator-new"></a>}}
\if{latex}{\out{\hypertarget{method-CurveGenerator-new}{}}}
\subsection{Method \code{new()}}{
Initialise class CurveGenerator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CurveGenerator$new(lengths)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lengths}}{vector of length classes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CurveGenerator-run"></a>}}
\if{latex}{\out{\hypertarget{method-CurveGenerator-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CurveGenerator$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CurveGenerator-generate_curve_df"></a>}}
\if{latex}{\out{\hypertarget{method-CurveGenerator-generate_curve_df}{}}}
\subsection{Method \code{generate_curve_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CurveGenerator$generate_curve_df()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CurveGenerator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CurveGenerator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CurveGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
