% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dome_from_logistic_generator.R
\name{DomeFromLogisticCurveGenerator}
\alias{DomeFromLogisticCurveGenerator}
\title{DomeFromLogisticCurveGenerator Class}
\description{
Class that creates a dome-shape curve ogive given the mode and spread
}
\section{Super class}{
\code{\link[utpbdomesel:CurveGenerator]{utpbdomesel::CurveGenerator}} -> \code{DomeFromLogisticCurveGenerator}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{k1}}{maximum retention length}

\item{\code{k2}}{curve spread}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DomeFromLogisticCurveGenerator-new}{\code{DomeFromLogisticCurveGenerator$new()}}
\item \href{#method-DomeFromLogisticCurveGenerator-run}{\code{DomeFromLogisticCurveGenerator$run()}}
\item \href{#method-DomeFromLogisticCurveGenerator-generate_curve_df}{\code{DomeFromLogisticCurveGenerator$generate_curve_df()}}
\item \href{#method-DomeFromLogisticCurveGenerator-clone}{\code{DomeFromLogisticCurveGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomeFromLogisticCurveGenerator-new"></a>}}
\if{latex}{\out{\hypertarget{method-DomeFromLogisticCurveGenerator-new}{}}}
\subsection{Method \code{new()}}{
Initialise class DomeFromLogisticCurveGenerator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomeFromLogisticCurveGenerator$new(lengths, k1, k2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lengths}}{vector of length classes}

\item{\code{k1}}{maximum retention length}

\item{\code{k2}}{curve spread}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomeFromLogisticCurveGenerator-run"></a>}}
\if{latex}{\out{\hypertarget{method-DomeFromLogisticCurveGenerator-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomeFromLogisticCurveGenerator$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomeFromLogisticCurveGenerator-generate_curve_df"></a>}}
\if{latex}{\out{\hypertarget{method-DomeFromLogisticCurveGenerator-generate_curve_df}{}}}
\subsection{Method \code{generate_curve_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomeFromLogisticCurveGenerator$generate_curve_df(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomeFromLogisticCurveGenerator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DomeFromLogisticCurveGenerator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomeFromLogisticCurveGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
