% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dome_from_logistic_generator.R
\name{DomeFromLogisticSelCurveGenerator}
\alias{DomeFromLogisticSelCurveGenerator}
\title{DomeFromLogisticSelCurveGenerator Class}
\description{
Class that creates a dome-shape curve off a logistic one
}
\section{Super class}{
\code{utpbdomesel::CurveGenerator} -> \code{DomeFromLogisticSelCurveGenerator}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{logistic_sfull}}{length at which a logistic function approaches 1 very rapidly}

\item{\code{dome_spread}}{sigma value of the new calculated dome-shape curve with max retention at logistic_sfull}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DomeFromLogisticSelCurveGenerator-new}{\code{DomeFromLogisticSelCurveGenerator$new()}}
\item \href{#method-DomeFromLogisticSelCurveGenerator-run}{\code{DomeFromLogisticSelCurveGenerator$run()}}
\item \href{#method-DomeFromLogisticSelCurveGenerator-generate_curve_df}{\code{DomeFromLogisticSelCurveGenerator$generate_curve_df()}}
\item \href{#method-DomeFromLogisticSelCurveGenerator-clone}{\code{DomeFromLogisticSelCurveGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomeFromLogisticSelCurveGenerator-new"></a>}}
\if{latex}{\out{\hypertarget{method-DomeFromLogisticSelCurveGenerator-new}{}}}
\subsection{Method \code{new()}}{
Initialise class DomeFromLogisticSelCurveGenerator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomeFromLogisticSelCurveGenerator$new(lengths, logistic_sfull)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lengths}}{vector of length classes}

\item{\code{logistic_sfull}}{length at which a logistic function approaches 1 very rapidly}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomeFromLogisticSelCurveGenerator-run"></a>}}
\if{latex}{\out{\hypertarget{method-DomeFromLogisticSelCurveGenerator-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomeFromLogisticSelCurveGenerator$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomeFromLogisticSelCurveGenerator-generate_curve_df"></a>}}
\if{latex}{\out{\hypertarget{method-DomeFromLogisticSelCurveGenerator-generate_curve_df}{}}}
\subsection{Method \code{generate_curve_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomeFromLogisticSelCurveGenerator$generate_curve_df()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomeFromLogisticSelCurveGenerator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DomeFromLogisticSelCurveGenerator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomeFromLogisticSelCurveGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
