% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittest_common_selectivity_finder.R
\name{FittestCommonDomeSelectivityFinder}
\alias{FittestCommonDomeSelectivityFinder}
\title{FittestCommonDomeSelectivityFinderClass}
\description{
Class that estimate the fittest selectivity curve through a number of trials
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catch_data}}{dataframe containing all raw data about gear and catches}

\item{\code{catch_context}}{R6 class holding information about catch_data columns and species' Linf}

\item{\code{haul_ids}}{list of haul id vectors}

\item{\code{mesh_sizes}}{name list of mesh sizes in ascending order}

\item{\code{rel_power}}{name list of mesh relative fishing power in ascending order}

\item{\code{sel_model}}{vector of models for which selectivity will be calculated}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{sel_model}}{vector of models for which selectivity will be calculated}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FittestCommonDomeSelectivityFinder-new}{\code{FittestCommonDomeSelectivityFinder$new()}}
\item \href{#method-FittestCommonDomeSelectivityFinder-run}{\code{FittestCommonDomeSelectivityFinder$run()}}
\item \href{#method-FittestCommonDomeSelectivityFinder-run_trials}{\code{FittestCommonDomeSelectivityFinder$run_trials()}}
\item \href{#method-FittestCommonDomeSelectivityFinder-fetch_sel_details}{\code{FittestCommonDomeSelectivityFinder$fetch_sel_details()}}
\item \href{#method-FittestCommonDomeSelectivityFinder-clone}{\code{FittestCommonDomeSelectivityFinder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FittestCommonDomeSelectivityFinder-new"></a>}}
\if{latex}{\out{\hypertarget{method-FittestCommonDomeSelectivityFinder-new}{}}}
\subsection{Method \code{new()}}{
Initialise class FittestCommonDomeSelectivityFinder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FittestCommonDomeSelectivityFinder$new(
  catch_data,
  catch_context,
  haul_ids,
  mesh_sizes,
  rel_power,
  sel_models
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{catch_data}}{a dataframe with the catch data}

\item{\code{catch_context}}{a R6 class with the context required to understand the catch data}

\item{\code{haul_ids}}{a list of haul id vectors from which to estimate a curve on each trial}

\item{\code{mesh_sizes}}{a named list with size of meshes involved}

\item{\code{rel_power}}{a named list with the relative fishing power of each mess}

\item{\code{sel_models}}{a character vector with the name of the models to estimate selectivity curves}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FittestCommonDomeSelectivityFinder-run"></a>}}
\if{latex}{\out{\hypertarget{method-FittestCommonDomeSelectivityFinder-run}{}}}
\subsection{Method \code{run()}}{
Run num_trials times the selectivity estimate and find the fittest one
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FittestCommonDomeSelectivityFinder$run(num_trials = 50)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_trials}}{number of trials}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FittestCommonDomeSelectivityFinder-run_trials"></a>}}
\if{latex}{\out{\hypertarget{method-FittestCommonDomeSelectivityFinder-run_trials}{}}}
\subsection{Method \code{run_trials()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FittestCommonDomeSelectivityFinder$run_trials(num_trials = 50)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FittestCommonDomeSelectivityFinder-fetch_sel_details"></a>}}
\if{latex}{\out{\hypertarget{method-FittestCommonDomeSelectivityFinder-fetch_sel_details}{}}}
\subsection{Method \code{fetch_sel_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FittestCommonDomeSelectivityFinder$fetch_sel_details(sel_detail_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FittestCommonDomeSelectivityFinder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FittestCommonDomeSelectivityFinder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FittestCommonDomeSelectivityFinder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
