% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MixinUtilities}
\alias{MixinUtilities}
\title{Mixin Utility class}
\value{
the backtransformed mode

the backtransformed standard deviation
}
\description{
It holds a few util methods used across all classes in the this R package
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MixinUtilities-create_empty_dataframe}{\code{MixinUtilities$create_empty_dataframe()}}
\item \href{#method-MixinUtilities-df_to_unname_matrix}{\code{MixinUtilities$df_to_unname_matrix()}}
\item \href{#method-MixinUtilities-back_transformed_mode}{\code{MixinUtilities$back_transformed_mode()}}
\item \href{#method-MixinUtilities-back_transformed_sd}{\code{MixinUtilities$back_transformed_sd()}}
\item \href{#method-MixinUtilities-clone}{\code{MixinUtilities$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixinUtilities-create_empty_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-MixinUtilities-create_empty_dataframe}{}}}
\subsection{Method \code{create_empty_dataframe()}}{
Create an empty dataframe
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixinUtilities$create_empty_dataframe(col_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col_names}}{vector with the column names of the empty dataframe}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixinUtilities-df_to_unname_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-MixinUtilities-df_to_unname_matrix}{}}}
\subsection{Method \code{df_to_unname_matrix()}}{
Convert a dataframe to matrix without row and column names. Such conversion could only be extensible
to a subset of columns only.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixinUtilities$df_to_unname_matrix(data, from_col = NULL, to_col = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col_names}}{vector with the column names of the empty dataframe}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixinUtilities-back_transformed_mode"></a>}}
\if{latex}{\out{\hypertarget{method-MixinUtilities-back_transformed_mode}{}}}
\subsection{Method \code{back_transformed_mode()}}{
Backtransform the log(mode(x)) into mode(x) applying e^(u - sigma^2)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixinUtilities$back_transformed_mode(log_mode, log_sigma, rel_mesh_size)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{log_mean}}{lognormal mean}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixinUtilities-back_transformed_sd"></a>}}
\if{latex}{\out{\hypertarget{method-MixinUtilities-back_transformed_sd}{}}}
\subsection{Method \code{back_transformed_sd()}}{
Backtransform the log(sqrt(sigma^2)) into sigma applying e^(2* u - sigma^2) * (e^(sigma^2) - 1)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixinUtilities$back_transformed_sd(log_mode, log_sigma)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{log_mean}}{lognormal mean}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MixinUtilities-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MixinUtilities-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MixinUtilities$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
