% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_logit_approximation.R
\name{NormalFromLogisticApproximator}
\alias{NormalFromLogisticApproximator}
\title{DNormalFromLogisticApproximator Class}
\description{
Class that finds the best normal distribution fit given a two-tail logistic functions stuck together and its s100.
This piece of code is a version from https://stats.stackexchange.com/a/630701/355585
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sl50}}{length at which fish could be caught with a 50\% chance}

\item{\code{sl95}}{length at which fish could be caught with a 90\% chance}

\item{\code{k1}}{mode parameter as described by the SELECT method}

\item{\code{length_increment}}{increments along the x axis for the length classes}

\item{\code{k2_scanned_interval}}{interval from which values are taken to optimise k2}

\item{\code{length_classes}}{vector with the length classes}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NormalFromLogisticApproximator-new}{\code{NormalFromLogisticApproximator$new()}}
\item \href{#method-NormalFromLogisticApproximator-run}{\code{NormalFromLogisticApproximator$run()}}
\item \href{#method-NormalFromLogisticApproximator-clone}{\code{NormalFromLogisticApproximator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NormalFromLogisticApproximator-new"></a>}}
\if{latex}{\out{\hypertarget{method-NormalFromLogisticApproximator-new}{}}}
\subsection{Method \code{new()}}{
Initialise class NormalFromLogisticApproximator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalFromLogisticApproximator$new(
  sl50,
  sl95,
  k1,
  length_increment,
  length_classes,
  k2_scanned_interval = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sl50}}{length at which fish could be caught with a 50\% chance}

\item{\code{sl95}}{length at which fish could be caught with a 90\% chance}

\item{\code{k1}}{mode parameter as described by the SELECT method}

\item{\code{length_increment}}{length of increments used in the k2 optimisation}

\item{\code{length_classes}}{vector with the length classes}

\item{\code{k2_scanned_interval}}{interval from which values are taken to optimise k2}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NormalFromLogisticApproximator-run"></a>}}
\if{latex}{\out{\hypertarget{method-NormalFromLogisticApproximator-run}{}}}
\subsection{Method \code{run()}}{
Find the fittest dome-shape curve that approximates another dome curve built from a two-parameter logistic functio
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalFromLogisticApproximator$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NormalFromLogisticApproximator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NormalFromLogisticApproximator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalFromLogisticApproximator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
