% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi_dome_from_logistic_generator.R
\name{QuasiDomeFromLogisticGenerator}
\alias{QuasiDomeFromLogisticGenerator}
\title{QuasiDomeFromLogisticGenerator Class}
\description{
Class that builds a quase-dome shape where the left limb is built by using a two-parameter logistic
formula and for the right one, the first maximun retention length class from the left limb logistic funciton
and the spread of a symmetric dome-sahpe. Such as spread is from a dome curve that was approximated using
the same logistic function as here.
}
\section{Super class}{
\code{\link[utpbdomesel:CurveGenerator]{utpbdomesel::CurveGenerator}} -> \code{QuasiDomeFromLogisticGenerator}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lengths}}{vector of length classes}

\item{\code{sl50}}{length at which fish could be caught with a 50\% chance}

\item{\code{sl95}}{length at which fish could be caught with a 90\% chance}

\item{\code{k2}}{right-limb curve spread}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QuasiDomeFromLogisticGenerator-new}{\code{QuasiDomeFromLogisticGenerator$new()}}
\item \href{#method-QuasiDomeFromLogisticGenerator-run}{\code{QuasiDomeFromLogisticGenerator$run()}}
\item \href{#method-QuasiDomeFromLogisticGenerator-generate_curve_df}{\code{QuasiDomeFromLogisticGenerator$generate_curve_df()}}
\item \href{#method-QuasiDomeFromLogisticGenerator-clone}{\code{QuasiDomeFromLogisticGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuasiDomeFromLogisticGenerator-new"></a>}}
\if{latex}{\out{\hypertarget{method-QuasiDomeFromLogisticGenerator-new}{}}}
\subsection{Method \code{new()}}{
Initialise class QuasiDomeFromLogisticGenerator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuasiDomeFromLogisticGenerator$new(lengths, sl50, sl95, k2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lengths}}{vector with the length classes}

\item{\code{sl50}}{length at which fish could be caught with a 50\% chance}

\item{\code{sl95}}{length at which fish could be caught with a 90\% chance}

\item{\code{k2}}{right-limb curve spread}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuasiDomeFromLogisticGenerator-run"></a>}}
\if{latex}{\out{\hypertarget{method-QuasiDomeFromLogisticGenerator-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuasiDomeFromLogisticGenerator$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuasiDomeFromLogisticGenerator-generate_curve_df"></a>}}
\if{latex}{\out{\hypertarget{method-QuasiDomeFromLogisticGenerator-generate_curve_df}{}}}
\subsection{Method \code{generate_curve_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuasiDomeFromLogisticGenerator$generate_curve_df(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QuasiDomeFromLogisticGenerator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QuasiDomeFromLogisticGenerator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QuasiDomeFromLogisticGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
