% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectivity.R
\name{SelectivityCurveGenerator}
\alias{SelectivityCurveGenerator}
\title{SelectivityCurveGeneratorr Class}
\description{
Wrapper class that call the millar SELECT method, two parameter logistic funtion and normal fixed function
to generate selection curves as requested
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SelectivityCurveGenerator-select_millar}{\code{SelectivityCurveGenerator$select_millar()}}
\item \href{#method-SelectivityCurveGenerator-two_params_logistic}{\code{SelectivityCurveGenerator$two_params_logistic()}}
\item \href{#method-SelectivityCurveGenerator-dome_approach_to_logistic}{\code{SelectivityCurveGenerator$dome_approach_to_logistic()}}
\item \href{#method-SelectivityCurveGenerator-clone}{\code{SelectivityCurveGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SelectivityCurveGenerator-select_millar"></a>}}
\if{latex}{\out{\hypertarget{method-SelectivityCurveGenerator-select_millar}{}}}
\subsection{Method \code{select_millar()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SelectivityCurveGenerator$select_millar(
  catch_data,
  catch_context,
  haul_ids,
  mesh_sizes,
  rel_power,
  sel_models,
  num_trials
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SelectivityCurveGenerator-two_params_logistic"></a>}}
\if{latex}{\out{\hypertarget{method-SelectivityCurveGenerator-two_params_logistic}{}}}
\subsection{Method \code{two_params_logistic()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SelectivityCurveGenerator$two_params_logistic()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SelectivityCurveGenerator-dome_approach_to_logistic"></a>}}
\if{latex}{\out{\hypertarget{method-SelectivityCurveGenerator-dome_approach_to_logistic}{}}}
\subsection{Method \code{dome_approach_to_logistic()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SelectivityCurveGenerator$dome_approach_to_logistic(lengths, logistic_sfull)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SelectivityCurveGenerator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SelectivityCurveGenerator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SelectivityCurveGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
