% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_generator.R
\name{TwoParamsLogisticSelCurveCalculator}
\alias{TwoParamsLogisticSelCurveCalculator}
\title{TwoParamsLogisticSelCurveCalculator Class}
\description{
Class that generates a logistic ogive from SL50 and SL95 parameters
}
\section{Super class}{
\code{utpbdomesel::CurveGenerator} -> \code{TwoParamsLogisticSelCurveCalculator}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sl50}}{length at which a logistic function reaches 50\%}

\item{\code{sl95}}{length at which a logistic function reaches 95\%}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TwoParamsLogisticSelCurveCalculator-new}{\code{TwoParamsLogisticSelCurveCalculator$new()}}
\item \href{#method-TwoParamsLogisticSelCurveCalculator-run}{\code{TwoParamsLogisticSelCurveCalculator$run()}}
\item \href{#method-TwoParamsLogisticSelCurveCalculator-generate_curve_df}{\code{TwoParamsLogisticSelCurveCalculator$generate_curve_df()}}
\item \href{#method-TwoParamsLogisticSelCurveCalculator-clone}{\code{TwoParamsLogisticSelCurveCalculator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TwoParamsLogisticSelCurveCalculator-new"></a>}}
\if{latex}{\out{\hypertarget{method-TwoParamsLogisticSelCurveCalculator-new}{}}}
\subsection{Method \code{new()}}{
Initialise class TwoParamsLogisticSelCurveCalculator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoParamsLogisticSelCurveCalculator$new(lengths, sl50, sl95)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lengths}}{vector of length classes}

\item{\code{sl50}}{length at which a logistic function reaches 50\%}

\item{\code{sl95}}{length at which a logistic function reaches 95\%}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TwoParamsLogisticSelCurveCalculator-run"></a>}}
\if{latex}{\out{\hypertarget{method-TwoParamsLogisticSelCurveCalculator-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoParamsLogisticSelCurveCalculator$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TwoParamsLogisticSelCurveCalculator-generate_curve_df"></a>}}
\if{latex}{\out{\hypertarget{method-TwoParamsLogisticSelCurveCalculator-generate_curve_df}{}}}
\subsection{Method \code{generate_curve_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoParamsLogisticSelCurveCalculator$generate_curve_df()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TwoParamsLogisticSelCurveCalculator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TwoParamsLogisticSelCurveCalculator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoParamsLogisticSelCurveCalculator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
