% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_generator.R
\name{TwoParLogisticCurveGenerator}
\alias{TwoParLogisticCurveGenerator}
\title{TwoParLogisticCurveGenerator Class}
\description{
Class that generates a logistic ogive from SL50 and SL95 parameters
}
\section{Super class}{
\code{\link[utpbdomesel:CurveGenerator]{utpbdomesel::CurveGenerator}} -> \code{TwoParLogisticCurveGenerator}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sl50}}{length at which a logistic function reaches 50\%}

\item{\code{sl95}}{length at which a logistic function reaches 95\%}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TwoParLogisticCurveGenerator-new}{\code{TwoParLogisticCurveGenerator$new()}}
\item \href{#method-TwoParLogisticCurveGenerator-run}{\code{TwoParLogisticCurveGenerator$run()}}
\item \href{#method-TwoParLogisticCurveGenerator-generate_curve_df}{\code{TwoParLogisticCurveGenerator$generate_curve_df()}}
\item \href{#method-TwoParLogisticCurveGenerator-clone}{\code{TwoParLogisticCurveGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TwoParLogisticCurveGenerator-new"></a>}}
\if{latex}{\out{\hypertarget{method-TwoParLogisticCurveGenerator-new}{}}}
\subsection{Method \code{new()}}{
Initialise class TwoParLogisticCurveGenerator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoParLogisticCurveGenerator$new(lengths, sl50, sl95)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lengths}}{vector of length classes}

\item{\code{sl50}}{length at which a logistic function reaches 50\%}

\item{\code{sl95}}{length at which a logistic function reaches 95\%}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TwoParLogisticCurveGenerator-run"></a>}}
\if{latex}{\out{\hypertarget{method-TwoParLogisticCurveGenerator-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoParLogisticCurveGenerator$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TwoParLogisticCurveGenerator-generate_curve_df"></a>}}
\if{latex}{\out{\hypertarget{method-TwoParLogisticCurveGenerator-generate_curve_df}{}}}
\subsection{Method \code{generate_curve_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoParLogisticCurveGenerator$generate_curve_df(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TwoParLogisticCurveGenerator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TwoParLogisticCurveGenerator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoParLogisticCurveGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
