% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dome_selectivity_estimator.R
\name{UtpbDomeSELECTEstimator}
\alias{UtpbDomeSELECTEstimator}
\title{DomeSelectivityEstimator Class}
\value{
a named list with SELECT object result

the SELECT object result
}
\description{
Class that estimate some shape selectivity for a list of gears using SELECT
}
\section{Super class}{
\code{\link[utpbdomesel:MixinUtilities]{utpbdomesel::MixinUtilities}} -> \code{UtpbDomeSELECTEstimator}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{length_fq}}{long dataframe with lengths and gears}

\item{\code{fishing_power}}{named list with the relative fishing power of each gear}

\item{\code{mesh_sizes}}{named list with every mesh size sorted by asceding order}

\item{\code{catch_context}}{R6 object with the context of the catch dataframe}

\item{\code{models}}{vector of model names to be estimated by SELECT}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UtpbDomeSELECTEstimator-new}{\code{UtpbDomeSELECTEstimator$new()}}
\item \href{#method-UtpbDomeSELECTEstimator-estimate_selectivity_by_models}{\code{UtpbDomeSELECTEstimator$estimate_selectivity_by_models()}}
\item \href{#method-UtpbDomeSELECTEstimator-estimate_selectivity_curve}{\code{UtpbDomeSELECTEstimator$estimate_selectivity_curve()}}
\item \href{#method-UtpbDomeSELECTEstimator-clone}{\code{UtpbDomeSELECTEstimator$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="utpbdomesel" data-topic="MixinUtilities" data-id="create_empty_dataframe"><a href='../../utpbdomesel/html/MixinUtilities.html#method-MixinUtilities-create_empty_dataframe'><code>utpbdomesel::MixinUtilities$create_empty_dataframe()</code></a></span></li>
<li><span class="pkg-link" data-pkg="utpbdomesel" data-topic="MixinUtilities" data-id="df_to_unname_matrix"><a href='../../utpbdomesel/html/MixinUtilities.html#method-MixinUtilities-df_to_unname_matrix'><code>utpbdomesel::MixinUtilities$df_to_unname_matrix()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtpbDomeSELECTEstimator-new"></a>}}
\if{latex}{\out{\hypertarget{method-UtpbDomeSELECTEstimator-new}{}}}
\subsection{Method \code{new()}}{
Initialise class DomeSelectivityEstimator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtpbDomeSELECTEstimator$new(
  length_fq,
  fishing_power,
  mesh_sizes,
  catch_context,
  models
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{length_fq}}{long dataframe with lengths and gears}

\item{\code{length_fq}}{long dataframe with lengths and gears}

\item{\code{fishing_power}}{named list with the relative fishing power of each gear}

\item{\code{mesh_sizes}}{named list with every mesh size}

\item{\code{catch_context}}{catch_context R6 object with the context of the catch dataframe}

\item{\code{models}}{to be estimated as part of the selectivity indirect methods}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtpbDomeSELECTEstimator-estimate_selectivity_by_models"></a>}}
\if{latex}{\out{\hypertarget{method-UtpbDomeSELECTEstimator-estimate_selectivity_by_models}{}}}
\subsection{Method \code{estimate_selectivity_by_models()}}{
Estimate the selectivity curve for a set of gears and for a specific model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtpbDomeSELECTEstimator$estimate_selectivity_by_models(
  min_padding = 0,
  up_to_linf = TRUE,
  plot = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{min_padding}}{padding that is substracted to the minimum length class}

\item{\code{up_to_linf}}{boolean flag indicating whether the length classed should be capped}

\item{\code{plot}}{boolean flag indicating whether the estimated curves and their deviations should be plot or not}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtpbDomeSELECTEstimator-estimate_selectivity_curve"></a>}}
\if{latex}{\out{\hypertarget{method-UtpbDomeSELECTEstimator-estimate_selectivity_curve}{}}}
\subsection{Method \code{estimate_selectivity_curve()}}{
Estimate the selectivity curve for a set of gears and for a specific model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtpbDomeSELECTEstimator$estimate_selectivity_curve(
  model,
  fishing_power,
  gear_catch,
  mesh_size,
  midpoints,
  plot = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{model used to estimate the curve}

\item{\code{fishing_power}}{a vectorwith the relative fishing power of gears involved in ascendng order by mesh size}

\item{\code{gear_catch}}{matrix of catches by gears (same number of rows as length classes)}

\item{\code{mesh_size}}{a vector with the mesh size in ascending order by mesh size}

\item{\code{midpoints}}{a vector with all length classes}

\item{\code{plot}}{boolean flag indicating whether the estimated curve and its deviations should be plot or not}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UtpbDomeSELECTEstimator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UtpbDomeSELECTEstimator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UtpbDomeSELECTEstimator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
